Refactor wizard to new modules
