+++ src/components/NewDraftWizard/WizardComponents/WizardHeader.js
@@ -1,0 +1,29 @@
+//@flow
+import React from "react"
+import Typography from "@material-ui/core/Typography"
+import { makeStyles } from "@material-ui/core/styles"
+
+const useStyles = makeStyles({
+  paperTitle: {
+    fontWeight: "bold",
+  },
+})
+
+type DraftHeaderProps = {
+  headerText: string,
+}
+
+const WizardHeader = ({ headerText }: DraftHeaderProps) => {
+  const classes = useStyles()
+  return (
+    <Typography
+      component="h1"
+      variant="subtitle1"
+      className={classes.paperTitle}
+    >
+      {headerText}
+    </Typography>
+  )
+}
+
+export default WizardHeader
