+++ src/components/NewDraftWizard/WizardSteps/WizardCreateFolderStep.js
@@ -3,14 +3,13 @@ import React from "react"
 import WizardHeader from "../WizardComponents/WizardHeader"
 import WizardSteps from "../WizardComponents/WizardSteps"
 
-const WizardShowSummaryStep = () => {
+const WizardCreateFolderStep = () => {
   return (
     <>
       <WizardHeader headerText="Create new draft folder" />
       <WizardSteps />
-      <WizardHeader headerText="Summary" />
     </>
   )
 }
 
-export default WizardShowSummaryStep
+export default WizardCreateFolderStep
