+++ src/features/wizardStepSlice.js
@@ -1,0 +1,15 @@
+import { createSlice } from "@reduxjs/toolkit"
+
+const initialState = -1
+
+const wizardStepSlice = createSlice({
+  name: "wizardStep",
+  initialState: initialState,
+  reducers: {
+    increment: state => state + 1,
+    decrement: state => state - 1,
+    reset: () => initialState,
+  },
+})
+export const { increment, decrement, reset } = wizardStepSlice.actions
+export default wizardStepSlice.reducer
